--- client/src/components/appbars/MaybeShowNavBar.js
@@ -1,19 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { useLocation } from 'react-router-dom';
-
-export default function MaybeShowNavBar({children}){
-    const location = useLocation();
-    const [showNavBar, setShowNavBar] = useState(false);
-
-    useEffect(()=>{
-        if(location.pathname==='/map-edit')
-            setShowNavBar(false);
-        else{
-            setShowNavBar(true);
-        }
-    },[location])
-
-    return (
-        <div>{showNavBar && children}</div>
-    )
-}
