+++ client/src/components/cards/index.js
@@ -1,0 +1,7 @@
+import DynamicCard from "./card/DynamicCard";
+import MapsCardSection from "./cardListFilter/MapsCardSection";
+import MapsPostsCardSection from "./cardListFilter/MapsPostsCardSelection";
+import PostsCardSection from "./cardListFilter/PostsCardSelection";
+import UsersCardSection from "./cardListFilter/UsersCardSelection";
+
+export { DynamicCard, MapsCardSection, MapsPostsCardSection, PostsCardSection, UsersCardSection }
