+++ client/src/components/modals/index.js
@@ -1,0 +1,8 @@
+import DeleteMapModal from './DeleteMapModal';
+import DeletePostModal from './DeletePostModal';
+import DeleteAccountModal from './DeleteAccountModal';
+import DeleteCommentModal from './DeleteCommentModal';
+import DeleteSubCommModal from './DeleteSubCommModal';
+import UploadMapErrorModal from './UploadMapErrorModal';
+
+export { DeleteMapModal, DeletePostModal, DeleteAccountModal, DeleteCommentModal, DeleteSubCommModal, UploadMapErrorModal }
