+++ client/src/components/screens/auth/index.js
@@ -1,0 +1,8 @@
+import Login from './Login';
+import Register from './Register';
+import AuthScreen from './AuthScreen';
+import PasswordScreen from './PasswordScreen';
+import ForgotPassword from './password/ForgotPassword';
+import ResetPassword from './password/ResetPassword';
+
+export { Login, Register, AuthScreen, PasswordScreen, ForgotPassword, ResetPassword  }
