+++ client/src/components/screens/home/index.js
@@ -1,0 +1,6 @@
+import Hero from "./splash/Hero";
+import HomeNavCard from "./splash/HomeNavCard";
+import UserHomeScreen from "./user/UserHomeScreen";
+import HomeScreen from "./HomeScreen";
+
+export { Hero, HomeNavCard, UserHomeScreen, HomeScreen }
