+++ client/src/components/screens/home/splash/Hero.js
