+++ client/src/components/screens/map/detail/GeojsonMap.js
