+++ client/src/components/screens/map/index.js
@@ -1,0 +1,6 @@
+import MapScreen from './detail/MapScreen';
+import MapComment from './detail/MapComment';
+import MapDetailsScreen from './detail/MapDetailsScreen';
+import MapEditScreen from './edit/MapEditScreen';
+
+export { MapScreen, MapComment, MapDetailsScreen, MapEditScreen }
