+++ client/src/components/screens/post/detail/PostSubComment.js
@@ -1,12 +1,12 @@
 import React, { useContext, useEffect } from "react";
-import { PostContext } from "../../../contexts/post";
+import { PostContext } from "../../../../contexts/post";
 import { Box } from "@mui/material";
 import { Typography } from "@mui/material";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 
 import DeleteForeverOutlinedIcon from '@mui/icons-material/DeleteForeverOutlined';
-import { GlobalStoreContext } from "../../../contexts/store";
-import { DeleteSubCommModal } from "../../index";
+import { GlobalStoreContext } from "../../../../contexts/store";
+import { DeleteSubCommModal } from "../../../index";
 
 export default function PostSubComment({parentComment, subcomment}) {
     const { store } = useContext(GlobalStoreContext);
