+++ client/src/components/screens/post/index.js
@@ -1,0 +1,5 @@
+import PostEditScreen from './edit/PostEditScreen';
+import PostDetailScreen from './detail/PostDetailScreen';
+import PostComment from './detail/PostComment';
+
+export { PostEditScreen, PostDetailScreen, PostComment }
