+++ client/src/components/warnings/index.js
@@ -1,0 +1,3 @@
+import NoAccessWarning from "./NoAccessWarning";
+
+export { NoAccessWarning }
